refactor exit composite
